feat: changes that generating openapi feature as plugin
