+++ _examples/13openapi/gen.go
@@ -5,22 +5,15 @@ package main
 
 import (
 	"context"
-	"encoding/json"
-	"io"
+	"fmt"
 	"log"
-	"reflect"
-	"strconv"
 
 	"m/13openapi/design"
+	"m/13openapi/myplugins/gendoc"
 
-	"github.com/getkin/kin-openapi/openapi3"
-	"github.com/podhmo/apikit/pkg/emitfile"
 	"github.com/podhmo/apikit/pkg/emitgo"
-	"github.com/podhmo/apikit/resolve"
 	"github.com/podhmo/apikit/web"
 	genchi "github.com/podhmo/apikit/web/webgen/gen-chi"
-	reflectopenapi "github.com/podhmo/reflect-openapi"
-	reflectshape "github.com/podhmo/reflect-shape"
 )
 
 func main() {
@@ -33,8 +26,6 @@ func main() {
 // TODO: set 404-handler
 
 func run() (err error) {
-	ctx := context.Background()
-
 	emitter := emitgo.NewConfigFromRelativePath(design.ListArticle, "..").NewEmitter()
 	defer emitter.EmitWith(&err)
 
@@ -58,77 +49,13 @@ func run() (err error) {
 		return err
 	}
 
-	////////////////////////////////////////
-	// TODO: share shape-extractor
-	rc := reflectopenapi.Config{
-		SkipValidation: true,
-		StrictSchema:   true,
-		Extractor:      c.Resolver.UnsafeShapeExtractor(),
-		IsRequiredCheckFunction: func(tag reflect.StructTag) bool {
-			v, _ := strconv.ParseBool(tag.Get("required"))
-			return v
-		},
-		Selector: &MergeParamsSelector{resolver: c.Resolver},
+	// generate openapi doc via custom plugin
+	if err := g.IncludePlugin(g.RootPkg, gendoc.Options{
+		OutputFile: "docs/openapi.json",
+		Handlers:   g.Handlers,
+	}); err != nil {
+		return fmt.Errorf("on gendoc plugin: %w", err)
 	}
 
-	doc, err := rc.BuildDoc(ctx, func(m *reflectopenapi.Manager) {
-		for _, h := range g.Handlers {
-			analyzed := h.Analyzed
-			metadata := h.MetaData
-
-			m.RegisterFunc(h.RawFn).After(func(op *openapi3.Operation) {
-				// e.g. articleID -> articleId
-				for _, p := range op.Parameters {
-					if p.Value.In == "path" {
-						for _, binding := range analyzed.Bindings.Path {
-							if binding.Name == p.Value.Name {
-								p.Value.Name = binding.Var.Name
-							}
-						}
-					}
-				}
-				m.Doc.AddOperation(metadata.Path, metadata.Method, op)
-			})
-		}
-	})
-	if err != nil {
-		log.Printf("WARNING: generate doc is failured %+v", err)
-	}
-	emitter.FileEmitter.Register("docs/openapi.json", emitfile.EmitFunc(func(w io.Writer) error {
-		enc := json.NewEncoder(w)
-		enc.SetIndent("", "  ")
-		return enc.Encode(doc)
-	}))
-	////////////////////////////////////////
 	return nil
-}
-
-type MergeParamsSelector struct {
-	resolver *resolve.Resolver
-	reflectopenapi.FirstParamOutputSelector
-}
-
-func (s *MergeParamsSelector) useArglist() {
-}
-func (s *MergeParamsSelector) SelectInput(fn reflectshape.Function) reflectshape.Shape {
-	if len(fn.Params.Values) == 0 {
-		return nil
-	}
-	shape, info, err := resolve.StructFromShape(s.resolver, fn)
-	if err != nil {
-		panic(err) // xxx
-	}
-
-	// add openapi tag
-	if indices, ok := info.GroupedByKind[resolve.KindPrimitive]; ok {
-		for _, i := range indices {
-			shape.Tags[i] = reflect.StructTag(string(shape.Tags[i]) + ` openapi:"path"`)
-		}
-	}
-	if indices, ok := info.GroupedByKind[resolve.KindPrimitivePointer]; ok {
-		for _, i := range indices {
-			shape.Tags[i] = reflect.StructTag(string(shape.Tags[i]) + ` openapi:"query"`)
-		}
-	}
-	return shape
 }
