+++ _examples/13openapi/myplugins/gendoc/plugin.go
@@ -1,0 +1,136 @@
+package gendoc
+
+import (
+	"context"
+	"encoding/json"
+	"io"
+	"log"
+	"reflect"
+	"strconv"
+
+	"github.com/getkin/kin-openapi/openapi3"
+	"github.com/podhmo/apikit/code"
+	"github.com/podhmo/apikit/pkg/emitfile"
+	"github.com/podhmo/apikit/pkg/emitgo"
+	"github.com/podhmo/apikit/pkg/tinypkg"
+	"github.com/podhmo/apikit/plugins"
+	"github.com/podhmo/apikit/resolve"
+	genchi "github.com/podhmo/apikit/web/webgen/gen-chi"
+	reflectopenapi "github.com/podhmo/reflect-openapi"
+	reflectshape "github.com/podhmo/reflect-shape"
+)
+
+type Options struct {
+	OutputFile string
+	Handlers   []genchi.Handler
+}
+
+func (o Options) IncludeMe(pc *plugins.PluginContext, here *tinypkg.Package) error {
+	return IncludeMe(
+		pc.Config, pc.Resolver, pc.Emitter,
+		here,
+		o.OutputFile,
+		o.Handlers,
+	)
+}
+
+func IncludeMe(
+	config *code.Config, resolver *resolve.Resolver, emitter *emitgo.Emitter,
+	here *tinypkg.Package,
+	outputFile string,
+	handlers []genchi.Handler,
+) error {
+	ctx := context.TODO() // hmm
+	if outputFile == "" {
+		log.Println("output filename is empty, so saving at docs/openapi.json")
+		outputFile = "docs/openapi.json"
+	}
+
+	////////////////////////////////////////
+	// TODO: share shape-extractor
+	rc := reflectopenapi.Config{
+		SkipValidation: true,
+		StrictSchema:   true,
+		Extractor:      resolver.UnsafeShapeExtractor(),
+		IsRequiredCheckFunction: func(tag reflect.StructTag) bool {
+			v, _ := strconv.ParseBool(tag.Get("required"))
+			return v
+		},
+		Selector: &MergeParamsSelector{resolver: resolver},
+	}
+
+	doc, err := rc.BuildDoc(ctx, func(m *reflectopenapi.Manager) {
+		for _, h := range handlers {
+			analyzed := h.Analyzed
+			metadata := h.MetaData
+
+			m.RegisterFunc(h.RawFn).After(func(op *openapi3.Operation) {
+				// e.g. articleID -> articleId
+				for _, p := range op.Parameters {
+					if p.Value.In == "path" {
+						for _, binding := range analyzed.Bindings.Path {
+							if binding.Name == p.Value.Name {
+								p.Value.Name = binding.Var.Name
+							}
+						}
+					}
+				}
+				m.Doc.AddOperation(metadata.Path, metadata.Method, op)
+			})
+		}
+	})
+	if err != nil {
+		log.Printf("WARNING: generate doc is failured %+v", err)
+		return nil
+	}
+
+	emitter.FileEmitter.Register(outputFile, emitfile.EmitFunc(func(w io.Writer) error {
+		enc := json.NewEncoder(w)
+		enc.SetIndent("", "  ")
+		return enc.Encode(doc)
+	}))
+	return nil
+}
+
+// selector ////////////////////////////////////////
+//
+// func(data Data, xxID string) { ... }
+//
+// to
+//
+// struct {
+//  Data
+//  xxID string `openapi:"path"`
+// }
+//
+
+// MergeParamsSelector is the selector with merging function arguments as single struct
+type MergeParamsSelector struct {
+	resolver *resolve.Resolver
+	reflectopenapi.FirstParamOutputSelector
+}
+
+func (s *MergeParamsSelector) useArglist() {
+}
+func (s *MergeParamsSelector) SelectInput(fn reflectshape.Function) reflectshape.Shape {
+	if len(fn.Params.Values) == 0 {
+		return nil
+	}
+	shape, info, err := resolve.StructFromShape(s.resolver, fn)
+	if err != nil {
+		panic(err) // xxx
+	}
+
+	// add openapi tag
+	if indices, ok := info.GroupedByKind[resolve.KindPrimitive]; ok {
+		for _, i := range indices {
+			shape.Tags[i] = reflect.StructTag(string(shape.Tags[i]) + ` openapi:"path"`)
+		}
+	}
+	if indices, ok := info.GroupedByKind[resolve.KindPrimitivePointer]; ok {
+		for _, i := range indices {
+			shape.Tags[i] = reflect.StructTag(string(shape.Tags[i]) + ` openapi:"query"`)
+		}
+	}
+	return shape
+}
